fix: 동문수첩 리뷰 수정
