+++ apps/web/src/entities/alumni-contacts/config/query/alumniContactsKeys.ts
@@ -2,10 +2,8 @@ import { GetAlumniContactsQuery } from '../../types';
 
 export const alumniContactsQueryKeys = {
   all: ['alumni-contacts'] as const,
-  lists: () => [...alumniContactsQueryKeys.all, 'list'] as const,
   list: (params: GetAlumniContactsQuery) =>
-    [...alumniContactsQueryKeys.lists(), params] as const,
-
+    [...alumniContactsQueryKeys.all, 'list', params] as const,
   details: () => [...alumniContactsQueryKeys.all, 'detail'] as const,
   detail: (id: string) => [...alumniContactsQueryKeys.details(), id] as const,
 };
