+++ apps/web/src/entities/profile/index.ts
@@ -1,3 +1,5 @@
 export * from './api';
 export * from './config';
 export * from './types';
+
+export * from './ui/ActionButtons';
