+++ apps/web/src/widgets/alumni-contacts/index.ts
@@ -3,3 +3,5 @@ export {
   AlumniContactsListServerComponent,
   AlumniContactsListLoadingView,
 } from './ui';
+
+export * from './ui';
