+++ apps/web/src/widgets/profile/index.ts
@@ -1,0 +1,1 @@
+export * from './model/useProfilePage';
